Fixes #384: Refactor utils and correct file imports
