+++ src/Utils/index.js
@@ -1,0 +1,2 @@
+export { default as urls } from './urls';
+export { default as isProduction } from './isProduction';
