+++ src/Utils/isProduction.js
@@ -1,4 +1,6 @@
-export const isProduction = () => {
+let isProduction = () => {
   let domain = window.location.hostname;
   return domain.indexOf('.susi.ai') > 0;
 };
+
+export default isProduction;
