+++ src/components/Auth/ChangePassword/ChangePassword.react.js
@@ -10,7 +10,7 @@ import PasswordField from 'material-ui-password-field';
 import Cookies from 'universal-cookie';
 import injectTapEventPlugin from 'react-tap-event-plugin';
 
-import urls from '../../../utils/urls';
+import { urls } from '../../../Utils';
 
 const cookies = new Cookies();
 injectTapEventPlugin();
