+++ src/components/Footer/Footer.react.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import susi from '../../images/susi-logo.svg';
 import './Footer.css';
 import { Link } from 'react-router-dom';
-import urls from '../../utils/urls';
+import { urls } from '../../Utils';
 
 class Footer extends Component {
   constructor(props) {
