+++ src/components/Settings/Settings.react.js
@@ -42,11 +42,9 @@ import LockIcon from 'material-ui/svg-icons/action/lock';
 import MyDevices from 'material-ui/svg-icons/device/devices';
 import MobileIcon from 'material-ui/svg-icons/hardware/phone-android';
 
-import urls from '../../utils/urls';
-
 import 'antd/dist/antd.css';
 import './Settings.css';
-import { isProduction } from '../../utils/helperFunctions';
+import { urls, isProduction } from '../../Utils';
 
 const cookieDomain = isProduction() ? '.susi.ai' : '';
