+++ src/stores/UserPreferencesStore.js
@@ -1,7 +1,7 @@
 import ChatAppDispatcher from '../dispatcher/ChatAppDispatcher';
 import ChatConstants from '../constants/ChatConstants';
 import { EventEmitter } from 'events';
-import urls from '../utils/urls';
+import { urls } from '../Utils';
 
 let ActionTypes = ChatConstants.ActionTypes;
 let CHANGE_EVENT = 'change';
